Implement is_following for FollowTreeNode
Allows the FollowTree construction method to identify nodes
that is can point to the new node, and then have them cascade
into the tree. This structure allows the construction to only
consider nodes at the root of the tree (there is a
hypothetical pointer to the user at the top, but it is
implemented as a forest of trees).