Remove logic from twitter_access in favor of <module>/tasks
There was a lot of stuff being done to try and organize the
twitter_access module to be smart about the requests that it was
sending. I decided in favor of trimming the insight_apis module
down to just its core functionality of providing access to another
site, and instead allow the various tasks that use the apis to
be smart with how they use it. insight_apis does not need to know
how to bulk package requests for users, when the code grabbing
users for a follow_tree can know exactly what it needs to do, and
can be reasonably expected to do it efficiently, and track
responses that indicate a rate limit is in effect, so it will stop
making requests.