Queue manager can create workers but not kill them
I'm going to need to rewrite my own worker that does a "work once"
then checks a list to see if it's been killed, and loops. In this
way, there will be one manager-worker that reads from a worker-
create queue, and all the others read from a cycling worker-
destroy queue or something like that.